backend/requirements.txt
pdfplumber==0.10.3
python-docx==1.1.0
nltk==3.8.1
pandas==2.1.4
openpyxl==3.1.2
flask==3.0.0
Werkzeug==3.0.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
reportlab==4.0.7
flask-cors==4.0.0
textstat==0.7.3
requests==2.31.0
python-dateutil==2.8.2